+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - #30, #48 - DSCacheFactory integration
 - #49 - DS.bindOne($scope, prop, resourceName, id)
 - #50 - DS.bindAll($scope, prop, resourceName, query)
+- #52 - DS.update(resourceName, id, attrs[, options]) (different from DS.save())
 - #54 - Adding functionality to resources
 
 ##### 0.8.1 - 02 May 2014